+++ controller/Api/v1/questions_controller.js
@@ -6,7 +6,8 @@ const Options = require("../../../model/Options");
 module.exports.create = async function (req, res) {
   if (req.body.question == undefined || req.body.question == null) {
     return res.status(400).json({
-      message: "error creating question wrong input key name use 'question' as key",
+      message:
+        "error creating question wrong input key name use 'question' as key",
     });
   }
   // create new question in Question model
@@ -42,6 +43,12 @@ module.exports.view = async function (req, res) {
 
 //view question and options with selected question id
 module.exports.viewQuestionWithID = async function (req, res) {
+  let questionIsExist = await Questions.exists({ _id: req.params.id });
+  if (questionIsExist == null) {
+    return res.status(404).json({
+      message: "question does not exists",
+    });
+  }
   await Questions.findById(req.params.id)
     .populate("options")
     .then((question) => {
@@ -51,7 +58,7 @@ module.exports.viewQuestionWithID = async function (req, res) {
     })
     .catch((err) => {
       res.status(404).json({
-        message: `No Data found`,
+        message: `Error finding question`,
       });
     });
 };
@@ -60,54 +67,59 @@ module.exports.viewQuestionWithID = async function (req, res) {
 module.exports.findQuestionandCreateOptions = async function (req, res) {
   if (req.body.text == undefined || req.body.text == null) {
     return res.status(400).json({
-      message: "error creating option no input provided or wrong input key name use 'text' as key",
+      message:
+        "error creating option no input provided or wrong input key name use 'text' as key",
     });
   }
 
   try {
-    await Questions.findById(req.params.id).then(async (question)  => {
-      
+    await Questions.findById(req.params.id).then(async (question) => {
       await Options.create({
         text: req.body.text,
         question_id: question.id,
       }).then(async (option) => {
-        
-        await Options.findByIdAndUpdate(option.id, {link_to_vote: `http://localhost:8000/api/v1/options/${option.id}/add_vote`}, {new: true})
-        .then((updateOption) => {
-          
-          question.options.push(updateOption)
-          question.save()
+        await Options.findByIdAndUpdate(
+          option.id,
+          {
+            link_to_vote: `http://localhost:8000/api/v1/options/${option.id}/add_vote`,
+          },
+          { new: true }
+        ).then((updateOption) => {
+          question.options.push(updateOption);
+          question.save();
           return res.status(200).json({
             option: updateOption,
-            message: "option created successfully"
-          })
-        })
-      })
-    })   
-    
+            message: "option created successfully",
+          });
+        });
+      });
+    });
   } catch (error) {
-    return res.status(404).json({
-      data: `Error finding question:  ${error}`
-    })
+    return res.status(400).json({
+      data: `Error finding question`,
+    });
   }
-
-}
-
+};
 
 //delete the question and it's associated options
 module.exports.deleteQuestion = async function (req, res) {
-  try {
-    await Questions.findById(req.params.id).then((question) => {
-
-      Options.deleteMany({_id: {$in: question.options}})
-      Questions.deleteOne({_id: question._id})
-      return res.status(200).json({
-        message: "question deleted successfully"
-      })
-    })
-  } catch (error) {
-    return res.status(200).json({
-      message: "question deleted successfully"
-    })
+  let questionIsExist = await Questions.exists({ _id: req.params.id });
+  if (questionIsExist == null) {
+    return res.status(404).json({
+      message: "question does not exists",
+    });
   }
-}
+  try {
+    await Questions.findById(req.params.id).then(async (question) => {
+      await Options.deleteMany({ _id: { $in: question.options } });
+      await Questions.deleteOne({ _id: question._id });
+      return res.status(200).json({
+        message: "question deleted successfully",
+      });
+    });
+  } catch (error) {
+    return res.status(400).json({
+      message: `Error deleting question`,
+    });
+  }
+};
